backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.28.1
python-dotenv==1.1.0
pydantic-settings==2.9.1
PyJWT==2.10.1
pytest==8.3.5
